Bug: fixed edit product. Also moved lists to script folder
